renamed index.html to the welcomePage because that's the first page that the user is going to load into
renamed index.html to the welcomePage because that's the first page that the user is going to load into, I'm keeping the welcomePage and generalBlogPostPage seperate from each other because the welcome page will look slightly different from the general blog post pages. Ideally we would use a template to construct an HTML page for each blog catagory upon request. General blog post will act as the landing page for the user after they clicked on a catagory from the welcome page.